Refactored chat functionality into chat.js
